uctypeslib2: Module with more helper routines for uctypes.
This is second such module after 'uctypeslib2', and intended to be a
superset (and in no way replacement) of it. The "superset" functionality
is intended to be achieved by importing 'uctypeslib' into 'uctypeslib2',
but in this initial version they're still indpendent.

The functionality offered by uctypeslib2 so far is pprint() function,
which can dump a structure descriptor back to form using uctypes.*
type constants. Besides debugging usage, it's also useful to prepare
a "more human readable" description using helper macros in uctypeslib,
then possibly run uctypes.calc_offsets(), then dump it in a still
human-readable way, but more efficient to import and use immediately.