uasyncio: remove_reader: Handle socket not in objmap gracefully.
Use self.objmap.pop(), similarly to how it's done in remove_write().
With remove_writer(), we may have a situation that socket was never
added to poller at all (if non-blocking writes to it always succeeded
on the 1st try). With remove_reader(), the situation is different -
we may attempt remove socket from poller multiple times. Specifically,
StreamReader functions automatically issue IOReadDone in case EOF is
seen (this is done to free up resources as soon as possible, not
waiting until user fully closes the socket). Additionally,
remove_reader() is called in case POLLHUP/POLERR is returned from
poll(), which is important to do because these flags are "sticky"
(so, if user doesn't handle them properly (== aclose() the socket
at once), we'd have problems).

Note that remove_writer() has poller.unregister() call protected
with try/except OSError(ENOENT). But currently, MicroPython's
uselect.poll.unregister() doesn't raise this exception if its
argument is not registered with the poller. The try/except is
apparently left from times uasyncio was running on CPython too.

Until situation is clarified, leave remove_writer as is, but
don't add try/except here.