cpython-uctypes: Add hacky implementation of bytearray_at().
The idea of bytearray_at() is that it captures reference to memory, and
that memory can be muted after that (externally or via returned bytearray).
Here, it is effectively implemented as bytes_at().

It's not immediately clear how to implement bytearray_at() via ctypes, and
implementation here to just let examples/unix/ffi_example.py in MicroPython
distro to run with CPython.